refactor(vue): rename internal vue utils
